command/requirements.txt
streamlit
streamlit-toggle-switch
paho-mqtt==1.6.1